@@ -3,43 +3,42 @@
     "version": "1.0.0",
     "private": true,
     "dependencies": {
-      "@reduxjs/toolkit": "^1.6.0",
-      "axios": "^0.21.1",
-      "d3": "^7.0.0",
-      "react": "^17.0.2",
-      "react-dom": "^17.0.2",
-      "react-redux": "^7.2.4",
-      "react-router-dom": "^5.2.0",
-      "redux": "^4.1.0",
-      "redux-devtools-extension": "^2.13.9",
-      "redux-thunk": "^2.3.0",
-      "react-scripts": "^4.0.3"
+        "@reduxjs/toolkit": "^1.6.0",
+        "axios": "^0.21.4",
+        "d3": "^7.9.0",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-redux": "^7.2.9",
+        "react-router-dom": "^5.2.0",
+        "react-scripts": "^4.0.3",
+        "redux": "^4.2.1",
+        "redux-devtools-extension": "^2.13.9",
+        "redux-thunk": "^2.4.2"
     },
     "scripts": {
-      "start": "react-scripts start",
-      "build": "react-scripts build",
-      "test": "react-scripts test",
-      "eject": "react-scripts eject"
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject"
     },
     "eslintConfig": {
-      "extends": [
-        "react-app",
-        "react-app/jest"
-      ]
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
     },
     "browserslist": {
-      "production": [
-        ">0.2%",
-        "not dead",
-        "not op_mini all"
-      ],
-      "development": [
-        "last 1 chrome version",
-        "last 1 firefox version",
-        "last 1 safari version"
-      ]
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
     },
     "author": "Your Name",
     "license": "MIT"
-  }
-  
+}
