@@ -2,24 +2,27 @@ const express = require('express');
 const mongoose = require('mongoose');
 const dotenv = require('dotenv');
 
-dotenv.config();
+dotenv.config(); // Ensure this line is present
 
 const app = express();
 app.use(express.json());
 
-mongoose.connect(process.env.MONGODB_URI, { useNewUrlParser: true, useUnifiedTopology: true })
-  .then(() => console.log('MongoDB connected'))
-  .catch(err => console.error(err));
+mongoose.connect(process.env.MONGODB_URI, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true
+})
+.then(() => console.log('MongoDB connected'))
+.catch(err => console.error('MongoDB connection error:', err));
 
-// Import routes
-const userRoutes = require('./routes/user');
-const eventRoutes = require('./routes/event');
+const userRoutes = require('./routes/users'); // Corrected from './routes/user'
+const eventRoutes = require('./routes/events'); // Corrected from './routes/event'
 const feedbackRoutes = require('./routes/feedback');
+const profileRoutes = require('./routes/profile');
 
-// Use routes
 app.use('/api/users', userRoutes);
 app.use('/api/events', eventRoutes);
 app.use('/api/feedback', feedbackRoutes);
+app.use('/api/profile', profileRoutes);
 
 const PORT = process.env.PORT || 5000;
 app.listen(PORT, () => {
