@@ -1,55 +1,46 @@
-import { 
-    USER_LOADED, 
-    AUTH_ERROR, 
-    LOGIN_SUCCESS, 
-    LOGIN_FAIL, 
-    LOGOUT, 
-    REGISTER_SUCCESS, 
-    REGISTER_FAIL 
-  } from '../actions/types';
-  
-  const initialState = {
-    token: localStorage.getItem('token'),
-    isAuthenticated: null,
-    loading: true,
-    user: null,
-  };
-  
-  function authReducer(state = initialState, action) {
-    const { type, payload } = action;
-  
-    switch (type) {
-      case USER_LOADED:
-        return {
-          ...state,
-          isAuthenticated: true,
-          loading: false,
-          user: payload,
-        };
-      case REGISTER_SUCCESS:
-      case LOGIN_SUCCESS:
-        localStorage.setItem('token', payload.token);
-        return {
-          ...state,
-          ...payload,
-          isAuthenticated: true,
-          loading: false,
-        };
-      case REGISTER_FAIL:
-      case AUTH_ERROR:
-      case LOGIN_FAIL:
-      case LOGOUT:
-        localStorage.removeItem('token');
-        return {
-          ...state,
-          token: null,
-          isAuthenticated: false,
-          loading: false,
-        };
-      default:
-        return state;
-    }
+import { REGISTER_SUCCESS, REGISTER_FAIL, USER_LOADED, AUTH_ERROR, LOGIN_SUCCESS, LOGIN_FAIL, LOGOUT } from '../actions/types';
+
+const initialState = {
+  token: localStorage.getItem('token'),
+  isAuthenticated: null,
+  loading: true,
+  user: null
+};
+
+const authReducer = (state = initialState, action) => {
+  const { type, payload } = action;
+
+  switch (type) {
+    case USER_LOADED:
+      return {
+        ...state,
+        isAuthenticated: true,
+        loading: false,
+        user: payload
+      };
+    case REGISTER_SUCCESS:
+    case LOGIN_SUCCESS:
+      localStorage.setItem('token', payload.token);
+      return {
+        ...state,
+        ...payload,
+        isAuthenticated: true,
+        loading: false
+      };
+    case REGISTER_FAIL:
+    case AUTH_ERROR:
+    case LOGIN_FAIL:
+    case LOGOUT:
+      localStorage.removeItem('token');
+      return {
+        ...state,
+        token: null,
+        isAuthenticated: false,
+        loading: false
+      };
+    default:
+      return state;
   }
-  
-  export default authReducer;
-  
+};
+
+export default authReducer;
