@@ -1,21 +1,9 @@
-import { createStore, combineReducers, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import authReducer from './reducers/authReducer';
-import eventReducer from './reducers/eventReducer';
-import feedbackReducer from './reducers/feedbackReducer';
+import { configureStore } from '@reduxjs/toolkit';
+import rootReducer from './reducers';
 
-const rootReducer = combineReducers({
-  auth: authReducer,
-  events: eventReducer,
-  feedback: feedbackReducer,
+const store = configureStore({
+  reducer: rootReducer,
+  devTools: process.env.NODE_ENV !== 'production',
 });
-
-const middleware = [thunk];
-
-const store = createStore(
-  rootReducer,
-  composeWithDevTools(applyMiddleware(...middleware))
-);
 
 export default store;
